Fix: AlarmSchedulerImpl.kt
Cambiar putExtra(EXTRA_ALARM) a Int para evitar error al inyectar en pendingIntent

+++ app/src/main/java/com/ibc/procrastinapp/data/alarm/AlarmSchedulerImpl.kt
@@ -19,7 +19,7 @@ class AlarmSchedulerImpl(
         val intent = Intent(context, AlarmReceiver::class.java).apply {
             // Puedes pasar datos extra al BroadcastReceiver
             putExtra("EXTRA_MESSAGE", task.title)
-            putExtra("EXTRA_ALARM_ID", task.id)
+            putExtra("EXTRA_ALARM_ID", task.id.toInt())
             // Es importante que el action o los data sean únicos por PendingIntent
             // si quieres tener múltiples alarmas distintas.
             // O usar el id en el requestCode del PendingIntent.
